Rename get command to get-managed
