+++ pypanther/get_managed_rule.py
