+++ pypanther/main.py
@@ -8,7 +8,7 @@ from gql.transport.aiohttp import log as aiohttp_logger
 
 from pypanther import testing, upload
 from pypanther.custom_logging import setup_logging
-from pypanther.setup_subparsers import setup_get_rule_parser, setup_list_rules_parser
+from pypanther.setup_subparsers import setup_get_managed_rule_parser, setup_list_rules_parser
 from pypanther.vendor.panther_analysis_tool import util
 from pypanther.vendor.panther_analysis_tool.command import standard_args
 from pypanther.vendor.panther_analysis_tool.config import dynaconf_argparse_merge, setup_dynaconf
@@ -100,20 +100,20 @@ def setup_parser() -> argparse.ArgumentParser:
     )
     version_parser.set_defaults(func=version)
 
-    # Get command
-    get_parser = subparsers.add_parser(
-        "get",
+    # Get managed command
+    get_managed_parser = subparsers.add_parser(
+        "get-managed",
         help="Get the class associated with a specific Panther-managed id",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
-    get_parser.set_defaults(func=help_printer(get_parser))
-    get_subparsers = get_parser.add_subparsers()
-    get_rule_parser = get_subparsers.add_parser(
+    get_managed_parser.set_defaults(func=help_printer(get_managed_parser))
+    get_subparsers = get_managed_parser.add_subparsers()
+    get_managed_rule_parser = get_subparsers.add_parser(
         name="rule",
         help="Get the class associated with a specific Panther-managed rule by id",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
-    setup_get_rule_parser(get_rule_parser)
+    setup_get_managed_rule_parser(get_managed_rule_parser)
 
     # List command
     list_parser = subparsers.add_parser(
